refactor: Add Card component and update page.tsx to use it for rendering values
